user-history-service/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-dotenv==1.0.0
databases==0.7.0
sqlalchemy>=1.4.42,<1.5
httpx==0.24.0
psycopg2-binary==2.9.5
asyncpg==0.27.0 